examples/D1G1T4L-HUM4N5/Mindbot/requirements.txt
aiohttp
python-dotenv
livekit
pytz
